requirements.txt
aiofiles==24.1.0
anyio==4.9.0
arabic-reshaper==3.0.0
asgi-csrf==0.11
asgiref==3.8.1
asn1crypto==1.5.1
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
click-default-group==1.2.4
colorama==0.4.6
crispy-bootstrap4==2024.10
cryptography==44.0.2
cssselect2==0.8.0
datasette==0.65.1
datasette-dateutil==0.3
defusedxml==0.7.1
Django==5.2
django-crispy-forms==2.4
django-filter==25.1
djangorestframework==3.16.0
et_xmlfile==2.0.0
exceptiongroup==1.3.0
Faker==37.1.0
flexcache==0.3
flexparser==0.4
h11==0.16.0
html5lib==1.1
httpcore==1.0.9
httpx==0.28.1
hupper==1.12.1
idna==3.10
itsdangerous==2.2.0
janus==2.0.0
Jinja2==3.1.6
lxml==5.4.0
MarkupSafe==3.0.2
mergedeep==1.3.4
openpyxl==3.1.5
oscrypto==1.3.0
pillow==11.2.1
platformdirs==4.3.8
pluggy==1.6.0
psutil==7.0.0
psycopg2-binary==2.9.10
pycparser==2.22
pyHanko==0.26.0
pyhanko-certvalidator==0.26.8
pylibdmtx==0.1.10
PyMySQL==1.1.1
pypdf==5.4.0
python-barcode==0.15.1
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
qrcode==8.1
reportlab==4.4.0
requests==2.32.3
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
svglib==1.5.1
tinycss2==1.4.0
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.3.1
uritools==4.0.3
urllib3==2.4.0
uvicorn==0.34.3
webencodings==0.5.1
whitenoise==6.9.0
xhtml2pdf==0.2.17
